Documents/Private/Проги/zakup.one/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
elasticsearch==8.11.0
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0



